backend/requirements.txt
fastapi
uvicorn
pyserial
websocket
pydantic